api/requirements.txt
Flask==1.1.2
flask-cors==3.0.8
psycopg2-binary==2.8.5
gunicorn==20.0.4
